+++ android/src/main/kotlin/com/appactor/appactor_flutter/AppActorFlutterPlugin.kt
@@ -12,7 +12,7 @@ import com.appactor.plugin.events.PluginEventListener
 
 class AppActorFlutterPlugin : FlutterPlugin, MethodChannel.MethodCallHandler, ActivityAware {
     private lateinit var channel: MethodChannel
-    private val mainHandler = Handler(Looper.getMainLooper())
+    private val mainHandler by lazy { Handler(Looper.getMainLooper()) }
 
     override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
         channel = MethodChannel(binding.binaryMessenger, "appactor_flutter")